refactor: reduce codebase — remove shims/dead code, dedup shared logic
Codebase-reduction sweep on the IR branch (~645 fewer source lines, 3 modules
deleted, all 849 tests green, ruff clean). Tracked in REFACTOR.md.

Removed (shims / dead code):
- no-op LaTeXRenderer.register(), deprecated git_version.format_version,
  get_template callbacks-rejection, formal-letter alias, getLatex() alias
- unused SegmentContext/AssetMapping (+ Document.segments/assets),
  ConversionContext.runtime_common, GenerationStrategy.prefer_inputs,
  Document.slot_inclusions, empty core.fonts package
- ir.Alignment/from_short and ir.CitationMode/Cite.mode (no reader sets,
  no writer reads them)
- shadowed _discover_local_templates, unreachable loader block, dead
  devtools try/except
- unused texsmith.extensions registry (drifted from DEFAULT_MARKDOWN_EXTENSIONS,
  now the single source of truth)

Dedup / SSOT:
- shared writers._ir_queries (citation/footnote parsing + _find_image/_find_table)
  used by both LaTeX and Typst writers
- BaseFragment.build_config/inject concrete defaults; 10 fragments stop
  restating them
- TemplateContextHelpers mixin for article/book/letter (_coerce_string/_format_authors)
- merged core.html_utils into adapters.html_utils; deleted dead templates.text;
  unified _discover_template_variables in context_usage

Moved:
- core.conversion.glossary -> core.glossary (no conversion-layer dependency)
