+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and the project follows [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 
-- **Typed intermediate representation (IR).** Conversion now flows through an explicit three-stage pipeline — `readers/html` → `texsmith.ir` (a typed, backend-agnostic `Block`/`Inline` document model) → `writers/{latex,typst}` — instead of mutating a BeautifulSoup tree in place into LaTeX strings. The IR is semantic, not LaTeX-shaped (a table column carries `Alignment.RIGHT`, never `"lrX"`; the backend writer computes the preamble), which is what makes a second output backend possible without duplicating the conversion logic. New public surface: `texsmith.ir` (nodes, enums, `NodeVisitor`), `texsmith.readers.html.HtmlReader`, and `texsmith.writers.latex.LaTeXWriter` / `texsmith.writers.typst.TypstWriter`.
+- **Typed intermediate representation (IR).** Conversion now flows through an explicit three-stage pipeline — `readers/html` → `texsmith.ir` (a typed, backend-agnostic `Block`/`Inline` document model) → `writers/{latex,typst}` — instead of mutating a BeautifulSoup tree in place into LaTeX strings. The IR is semantic, not LaTeX-shaped (a table carries the validated `texsmith.extensions.tables` model with one-letter semantic alignment, never a `"lrX"` preamble; the backend writer computes the preamble), which is what makes a second output backend possible without duplicating the conversion logic. New public surface: `texsmith.ir` (nodes, enums, `NodeVisitor`), `texsmith.readers.html.HtmlReader`, and `texsmith.writers.latex.LaTeXWriter` / `texsmith.writers.typst.TypstWriter`.
 - **Typst output backend.** A new `--format {latex,typst}` flag (default `latex`) selects the output engine, so a single Markdown/HTML source can be built with a LaTeX engine (Tectonic, TeX Live, lualatex…) **or** with [Typst](https://typst.app/). The `article` and `book` templates are Typst-compatible: title, authors (with affiliations), subtitle, date/version, abstract, table of contents, section/heading numbering, native `#bibliography`, and book chapters. Templates declare a `[typst.template]` manifest section (parallel to `[latex.template]`) with a `template.typ` scaffolding; the Typst path routes through the template system, with a standalone fallback for the no-template case. The covered node subset is broad (text/formatting, headers, lists, code, math via `mitex`, links, images, tables, citations, footnotes, admonitions, definition lists); a few nodes (`MarginNote`, rich yaml/data tables) raise an explicit, localised `TypstWriteError` rather than degrading silently.
 - **Pip-installable Typst compiler.** A new `texsmith[typst]` optional extra bundles the Typst compiler (the PyPI `typst` package), so `pip install "texsmith[typst]"` / `uv tool install "texsmith[typst]"` enables `--build` for Typst with no system toolchain — analogous to the live Tectonic download. A system `typst` binary on `PATH` is used as a fallback when present. `.typ` emission never depends on a compiler being installed. (Note: the bundled PyPI compiler can lag recent Typst releases; for documents using recent `mitex` math, prefer a system `typst` binary.)
 - New documentation page `docs/guide/plumbing/backends.md` describing the IR pipeline, the `--format` flag, Typst installation paths, and the covered/uncovered Typst scope.
@@ -25,6 +25,8 @@ and the project follows [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - `escape_latex_chars` now lives in `texsmith.writers.latex.escaper` (escaping is a backend concern); `texsmith.adapters.latex.utils` re-exports it.
 - Pinned Tectonic version for the live download bumped `0.15.0` → `0.16.9` (release asset naming verified unchanged; the `--version` cache probe relies on a flag that was dropped in 0.16.0 and restored in 0.16.9).
 - Typst PDF compilation now **prefers the system `typst` binary** over the embedded PyPI compiler, falling back to the package when no binary is on `PATH`. The bundled package floats to the latest Typst (`>=0.15`), which drops symbols `mitex` still emits (e.g. `diff` for `\partial`); the user-installed binary tracks `mitex`'s supported versions, so preferring it keeps recent math compiling.
+- **De-duplicated shared logic across backends and fragments.** Citation/footnote-identifier parsing (`_normalise_footnote_id`, `_split_citation_keys`, `_citation_keys_from_payload`, the DOI/citation regexes) and the `_find_image` / `_find_table` IR lookups, previously copied verbatim in both the LaTeX and Typst writers, now live in one shared `texsmith.writers._ir_queries` module. `BaseFragment.build_config` / `inject` gained concrete defaults (`config_cls.from_context` / `config.inject_into`), so the ten fragments that followed that convention no longer each restate it (`frame` / `geometry` keep their bespoke parsing). The `_coerce_string` / `_format_authors` helpers copied across the article/book/letter templates are now a single `TemplateContextHelpers` mixin in `texsmith.templates.common`.
+- Merged `texsmith.core.html_utils` (`strip_html_comments`) into `texsmith.adapters.html_utils` and deleted the dead `texsmith.core.templates.text` module (the live `_squash_blank_lines` lives on in `wrapper`); `_discover_template_variables` is now defined once in `context_usage` and imported by `wrapper`. Front-matter glossary parsing moved from `texsmith.core.conversion.glossary` to `texsmith.core.glossary` (it has no conversion-layer dependency and sits beside the other front-matter resolvers).
 
 ### Removed
 
@@ -33,6 +35,8 @@ and the project follows [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - The obsolete `--debug-rules` flag and its rule-description plumbing (a vestige of the removed engine that emitted a warning on every conversion).
 - The `_FallbackConverter` placeholder-PDF shim.
 - The top-level shim modules `index.py`, `texlogos.py`, `progressbar.py`, `quotes.py`, and `smart_dashes.py`; their implementations now live under `texsmith.extensions.*` and the entry points point there directly.
+- **Codebase-reduction sweep (≈645 fewer source lines, 3 modules deleted).** Removed the remaining shims and dead code on the IR branch: the no-op `LaTeXRenderer.register()`, the deprecated `git_version.format_version` forwarder, the `get_template(..., callbacks=…)` rejection branch, the `formal-letter` template alias, and the camelCase `GeometryFragment.getLatex()` alias. Deleted unused machinery never read by the pipeline — the `SegmentContext` / `AssetMapping` dataclasses with `Document.segments` / `Document.assets`, `ConversionContext.runtime_common`, `GenerationStrategy.prefer_inputs`, `Document.slot_inclusions` (a duplicate of `slot_includes`); the empty `texsmith.core.fonts` package; and the unused `ir.Alignment` / `Alignment.from_short` and `ir.CitationMode` / `Cite.mode` (no reader sets them and no writer reads them). Removed a shadowed duplicate `_discover_local_templates`, an unreachable post-`return` block in the template loader, and a dead `try/except` in `devtools`.
+- The entirely unused Markdown-extension registry in `texsmith.extensions` (`ExtensionSpec`, `_EXTENSIONS`, `available_extensions`, `get_extension_spec`, `load_markdown_extension`, `load_mkdocs_plugin`) — it had no callers and had drifted out of sync with the authoritative `texsmith.adapters.markdown.DEFAULT_MARKDOWN_EXTENSIONS`, which is now the single source of truth.
 
 ## [0.3.1] - 2026-06-12
