--- src/texsmith/core/fonts/__init__.py
@@ -1,3 +1,0 @@
-"""Font utilities used by TexSmith's core rendering pipeline."""
-
-__all__ = []
