+++ src/texsmith/core/glossary.py
