+++ src/texsmith/extensions/__init__.py
@@ -1,150 +1,11 @@
-"""Central registry for TeXSmith's bundled Markdown extensions."""
+"""TeXSmith's bundled Python-Markdown extensions.
+
+Each submodule ships a ``markdown.Extension`` (and, for ``index``, an MkDocs
+plugin). The authoritative list of extensions enabled by the conversion
+pipeline lives in
+:data:`texsmith.adapters.markdown.DEFAULT_MARKDOWN_EXTENSIONS`; the same entries
+are also exposed to third-party Markdown/MkDocs setups via the entry points
+declared in ``pyproject.toml``.
+"""
 
 from __future__ import annotations
-
-from collections.abc import Callable
-from dataclasses import dataclass
-from importlib import import_module
-from typing import Any
-
-
-__all__ = [
-    "ExtensionSpec",
-    "available_extensions",
-    "get_extension_spec",
-    "load_markdown_extension",
-    "load_mkdocs_plugin",
-]
-
-
-def _load_attribute(path: str) -> Any:
-    module_name, _, attribute = path.partition(":")
-    if not module_name or not attribute:
-        msg = f"Extension entry point '{path}' must use the 'module:attribute' format."
-        raise ValueError(msg)
-    module = import_module(module_name)
-    target: Any = module
-    for chunk in attribute.split("."):
-        target = getattr(target, chunk)
-    return target
-
-
-def _normalise_slug(value: str) -> str:
-    slug = value.split(":", 1)[0].lower()
-    if slug.startswith("texsmith."):
-        return slug.removeprefix("texsmith.")
-    return slug
-
-
-@dataclass(frozen=True, slots=True)
-class ExtensionSpec:
-    """Describe how to import Markdown and renderer hooks for an extension."""
-
-    slug: str
-    markdown_entry: str
-    mkdocs_entry: str | None = None
-    description: str | None = None
-
-    @property
-    def package_name(self) -> str:
-        return f"texsmith.{self.slug}"
-
-
-_EXTENSIONS: dict[str, ExtensionSpec] = {
-    "index": ExtensionSpec(
-        slug="index",
-        markdown_entry="texsmith.extensions.index:TexsmithIndexExtension",
-        mkdocs_entry="texsmith.extensions.index:IndexPlugin",
-        description="Inline hashtags converted into LaTeX index entries.",
-    ),
-    "texlogos": ExtensionSpec(
-        slug="texlogos",
-        markdown_entry="texsmith.extensions.texlogos:TexLogosExtension",
-        description="Semantic spans for TeX logos rendered in Markdown and LaTeX.",
-    ),
-    "smallcaps": ExtensionSpec(
-        slug="smallcaps",
-        markdown_entry="texsmith.extensions.smallcaps:SmallCapsExtension",
-        description="Maps '__text__' syntax to small caps spans.",
-    ),
-    "latex_raw": ExtensionSpec(
-        slug="latex_raw",
-        markdown_entry="texsmith.extensions.latex_raw:LatexRawExtension",
-        description="Fence supporting inline raw LaTeX payloads.",
-    ),
-    "latex_text": ExtensionSpec(
-        slug="latex_text",
-        markdown_entry="texsmith.extensions.latex_text:LatexTextExtension",
-        description="Styles the literal 'LaTeX' token with a dedicated span.",
-    ),
-    "smart_dashes": ExtensionSpec(
-        slug="smart_dashes",
-        markdown_entry="texsmith.extensions.smart_dashes:TexsmithSmartDashesExtension",
-        description="Converts '--' and '---' into typographic dashes outside code.",
-    ),
-    "invisible_chars": ExtensionSpec(
-        slug="invisible_chars",
-        markdown_entry="texsmith.extensions.invisible_chars:InvisibleCharsExtension",
-        description="Replaces zero-width invisible characters (ZWSP, ZWNJ, ZWJ, BOM) with a non-breaking space.",
-    ),
-    "missing_footnotes": ExtensionSpec(
-        slug="missing_footnotes",
-        markdown_entry="texsmith.extensions.missing_footnotes:MissingFootnotesExtension",
-        description="Warns about references to undefined footnotes.",
-    ),
-    "multi_citations": ExtensionSpec(
-        slug="multi_citations",
-        markdown_entry="texsmith.extensions.multi_citations:MultiCitationExtension",
-        description="Normalises '^[a,b]' inline citations before footnotes run.",
-    ),
-    "mermaid": ExtensionSpec(
-        slug="mermaid",
-        markdown_entry="texsmith.extensions.mermaid:MermaidExtension",
-        description="Inlines Mermaid diagrams referenced via Markdown images.",
-    ),
-    "progressbar": ExtensionSpec(
-        slug="progressbar",
-        markdown_entry="texsmith.extensions.progressbar:ProgressBarExtension",
-        description="Renders `[=50%]` progress blocks via the LaTeX progressbar package.",
-    ),
-    "yaml_tables": ExtensionSpec(
-        slug="yaml_tables",
-        markdown_entry="texsmith.extensions.tables:YamlTableExtension",
-        description="YAML-described tables with multirow/multicolumn support.",
-    ),
-    "marginnote": ExtensionSpec(
-        slug="marginnote",
-        markdown_entry="texsmith.extensions.marginnote:MarginNoteExtension",
-        description="Inline `{margin}[note]{side?}` syntax rendered as LaTeX margin notes.",
-    ),
-}
-
-
-def available_extensions() -> list[ExtensionSpec]:
-    """Return the registered extension specs sorted by slug."""
-    return [_EXTENSIONS[key] for key in sorted(_EXTENSIONS)]
-
-
-def get_extension_spec(name: str) -> ExtensionSpec:
-    """Look up the runtime spec for a given extension slug or qualified name."""
-    slug = _normalise_slug(name)
-    try:
-        return _EXTENSIONS[slug]
-    except KeyError as exc:  # pragma: no cover - defensive
-        raise KeyError(f"No TeXSmith extension named '{name}'.") from exc
-
-
-def load_markdown_extension(name: str, **config: Any) -> Any:
-    """Instantiate a Python-Markdown extension by slug or qualified name."""
-    spec = get_extension_spec(name)
-    factory: Callable[..., Any] = _load_attribute(spec.markdown_entry)
-    return factory(**config)
-
-
-def load_mkdocs_plugin(name: str) -> type[Any] | None:
-    """Return the MkDocs plugin class for an extension, if any."""
-    spec = get_extension_spec(name)
-    if not spec.mkdocs_entry:
-        return None
-    plugin_cls: type[Any] = _load_attribute(spec.mkdocs_entry)
-    return plugin_cls
