+++ src/texsmith/fragments/callouts/__init__.py
@@ -51,21 +51,6 @@ class CalloutsFragment(BaseFragment[CalloutsConfig]):
     source: ClassVar[Path] = Path(__file__).with_name("ts-callouts.jinja.sty")
     context_defaults: ClassVar[dict[str, Any]] = {"extra_packages": ""}
 
-    def build_config(
-        self, context: Mapping[str, Any], overrides: Mapping[str, Any] | None = None
-    ) -> CalloutsConfig:
-        _ = overrides
-        return self.config_cls.from_context(context)
-
-    def inject(
-        self,
-        config: CalloutsConfig,
-        context: dict[str, Any],
-        overrides: Mapping[str, Any] | None = None,
-    ) -> None:
-        _ = overrides
-        config.inject_into(context)
-
     def should_render(self, config: CalloutsConfig) -> bool:
         return config.uses_callouts
