+++ src/texsmith/fragments/code/__init__.py
@@ -48,21 +48,6 @@ class CodeFragment(BaseFragment[CodeConfig]):
     source: ClassVar[Path] = Path(__file__).with_name("ts-code.jinja.sty")
     context_defaults: ClassVar[dict[str, Any]] = {"extra_packages": ""}
 
-    def build_config(
-        self, context: Mapping[str, Any], overrides: Mapping[str, Any] | None = None
-    ) -> CodeConfig:
-        _ = overrides
-        return self.config_cls.from_context(context)
-
-    def inject(
-        self,
-        config: CodeConfig,
-        context: dict[str, Any],
-        overrides: Mapping[str, Any] | None = None,
-    ) -> None:
-        _ = overrides
-        config.inject_into(context)
-
     def should_render(self, config: CodeConfig) -> bool:
         return config.uses_code
