+++ src/texsmith/fragments/todolist/__init__.py
@@ -37,21 +37,6 @@ class TodolistFragment(BaseFragment[TodolistConfig]):
     source: ClassVar[Path] = Path(__file__).with_name("ts-todolist.jinja.sty")
     context_defaults: ClassVar[dict[str, Any]] = {"extra_packages": ""}
 
-    def build_config(
-        self, context: Mapping[str, Any], overrides: Mapping[str, Any] | None = None
-    ) -> TodolistConfig:
-        _ = overrides
-        return self.config_cls.from_context(context)
-
-    def inject(
-        self,
-        config: TodolistConfig,
-        context: dict[str, Any],
-        overrides: Mapping[str, Any] | None = None,
-    ) -> None:
-        _ = overrides
-        config.inject_into(context)
-
     def should_render(self, config: TodolistConfig) -> bool:
         return config.uses_todolist
