+++ src/texsmith/fragments/typesetting/__init__.py
@@ -93,21 +93,6 @@ class TypesettingFragment(BaseFragment[TypesettingConfig]):
     source: ClassVar[Path] = Path(__file__).with_name("ts-typesetting.tex.jinja")
     context_defaults: ClassVar[dict[str, Any]] = {"extra_packages": ""}
 
-    def build_config(
-        self, context: Mapping[str, Any], overrides: Mapping[str, Any] | None = None
-    ) -> TypesettingConfig:
-        _ = overrides
-        return self.config_cls.from_context(context)
-
-    def inject(
-        self,
-        config: TypesettingConfig,
-        context: dict[str, Any],
-        overrides: Mapping[str, Any] | None = None,
-    ) -> None:
-        _ = overrides
-        config.inject_into(context)
-
     def should_render(self, config: TypesettingConfig) -> bool:
         return config.enabled()
