+++ tests/test_git_version.py
@@ -7,6 +7,7 @@ import warnings
 import pytest
 
 from texsmith.core import git_version
+from texsmith.core.document_version import format_version
 
 
 def _git(repo: Path, *args: str) -> None:
@@ -21,14 +22,14 @@ def _clear_cache() -> None:
 
 
 def test_format_version_returns_free_form_string_unchanged() -> None:
-    assert git_version.format_version("Draft 3") == "Draft 3"
-    assert git_version.format_version("  consolidated v2  ") == "consolidated v2"
+    assert format_version("Draft 3") == "Draft 3"
+    assert format_version("  consolidated v2  ") == "consolidated v2"
 
 
 def test_format_version_ignores_empty_input() -> None:
-    assert git_version.format_version("") == ""
-    assert git_version.format_version(None) == ""
-    assert git_version.format_version("   ") == ""
+    assert format_version("") == ""
+    assert format_version(None) == ""
+    assert format_version("   ") == ""
 
 
 def test_format_version_resolves_git_keyword(tmp_path: Path) -> None:
@@ -42,7 +43,7 @@ def test_format_version_resolves_git_keyword(tmp_path: Path) -> None:
     _git(repo, "commit", "-q", "-m", "init")
     _git(repo, "tag", "v1.0.0")
 
-    resolved = git_version.format_version("git", cwd=repo)
+    resolved = format_version("git", cwd=repo)
     assert resolved == "v1.0.0"
 
 
@@ -58,7 +59,7 @@ def test_format_version_marks_dirty_worktree(tmp_path: Path) -> None:
     _git(repo, "tag", "v0.1.0")
     (repo / "README.md").write_text("dirty", encoding="utf-8")
 
-    assert git_version.format_version("git", cwd=repo) == "v0.1.0-dirty"
+    assert format_version("git", cwd=repo) == "v0.1.0-dirty"
 
 
 def test_format_version_falls_back_to_short_hash(tmp_path: Path) -> None:
@@ -71,7 +72,7 @@ def test_format_version_falls_back_to_short_hash(tmp_path: Path) -> None:
     _git(repo, "add", ".")
     _git(repo, "commit", "-q", "-m", "init")
 
-    resolved = git_version.format_version("git", cwd=repo)
+    resolved = format_version("git", cwd=repo)
     assert resolved
     assert len(resolved) >= 7  # short hash, optionally with -dirty suffix
 
@@ -81,7 +82,7 @@ def test_format_version_without_repo_warns(tmp_path: Path) -> None:
     not_a_repo.mkdir()
     with warnings.catch_warnings(record=True) as captured:
         warnings.simplefilter("always")
-        result = git_version.format_version("git", cwd=not_a_repo)
+        result = format_version("git", cwd=not_a_repo)
     assert result == ""
     assert any("no git repository" in str(w.message) for w in captured)
 
@@ -97,5 +98,5 @@ def test_format_version_keyword_is_case_insensitive(tmp_path: Path) -> None:
     _git(repo, "commit", "-q", "-m", "init")
     _git(repo, "tag", "v2.0")
 
-    assert git_version.format_version("Git", cwd=repo) == "v2.0"
-    assert git_version.format_version("GIT", cwd=repo) == "v2.0"
+    assert format_version("Git", cwd=repo) == "v2.0"
+    assert format_version("GIT", cwd=repo) == "v2.0"
